TMDB Movies Project
The TMDB movie dataset contains information about 10,000 movies collected from The Movie Database (TMDB). It contains data on the plot, cast, crew, budget, user ratings and revenues of several films.
This dataset is analysed as a project to show the entire data analysis process from cleaning and wrangling to analysis to  exploration using different visualization charts. The analysis of this dataset is used to answer set questions and draw conclusions based on the relationship and dependency of variables.