A/B Testing
This project involves working to understand the results of an A/B test run by an e-commerce website.  The goal is to help the company understand if they should implement the new page, keep the old page, or perhaps run the experiment longer to make their decision. To establish the accuracy of the result obtained from the A/B test, a logistic regression model is created and a factor of the country a user lives in is also added to the model to measure its impact on the conversion rate.